src/requirements.txt
boto==2.39.0
dj-database-url==0.3.0
Django==1.9.1
django-account-manager==0.0.3
django-filter==0.12.0
django-herokuapp==0.9.20
django-require==1.0.9
django-require-s3==1.0.0
django-storages==1.1.8
djangorestframework==3.3.2
git==3.2.6
gunicorn==19.4.5
Markdown==2.6.5
psycopg2==2.6.1
pytz==2015.7
sh==1.11
SQLAlchemy==1.0.11
waitress==0.8.10
wheel==0.24.0
whitenoise==2.0.6
